+++ cryptographic-message-syntax/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 Released on ReleaseDate.
 
+* `SignerBuilder` gained a `new_with_signer_identifier()` that allows constructing
+  from a `SignerIdentifier` instead of a `CapturedX509Certificate`. This API allows
+  usage in alternate signing scenarios, such as those found in RFC 5272. Contributed
+  by Outurnate in #8.
+
 ## 0.21.0
 
 Released on 2023-01-21.